Added new ResponseBody POJO and made the UserPojo and UserSettingsPojo extend from it
Renamed the resendConfirmation endpoint to resendVerificationEmail
