+++ webofneeds/won-owner-webapp/src/main/java/won/owner/pojo/UserPojo.java
@@ -10,23 +10,11 @@ import org.hibernate.validator.constraints.NotEmpty;
  * User: t.kozel
  * Date: 11/12/13
  */
-public class UserPojo {
-
-	@NotEmpty
-	private String username;
-
-	@NotEmpty
+public class UserPojo extends UsernamePojo {
+    @NotEmpty
 	private String password;
 
     private boolean privateIdUser;
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(final String username) {
-		this.username = username;
-	}
 
 	public String getPassword() {
 		return password;
